# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DramaGen API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export type TranslateRequestMode =
  (typeof TranslateRequestMode)[keyof typeof TranslateRequestMode];

export const TranslateRequestMode = {
  rage: "rage",
  passive_aggressive: "passive_aggressive",
  disappointed_parent: "disappointed_parent",
  telenovela: "telenovela",
  tiktok: "tiktok",
  drill_sergeant: "drill_sergeant",
  corporate: "corporate",
  ice_cold: "ice_cold",
} as const;

export interface TranslateRequest {
  /** @maxLength 500 */
  text: string;
  mode: TranslateRequestMode;
}

export interface TranslateResponse {
  original: string;
  translated: string;
  mode: string;
}

export interface PresetVoice {
  id: string;
  voice_id?: string;
  name: string;
  description: string;
  emoji: string;
  ready: boolean;
}

export interface PresetVoicesResponse {
  voices: PresetVoice[];
}

export interface CloneVoiceResponse {
  voice_id: string;
  name: string;
}

export type GenerateRequestVoiceType =
  (typeof GenerateRequestVoiceType)[keyof typeof GenerateRequestVoiceType];

export const GenerateRequestVoiceType = {
  clone: "clone",
  preset: "preset",
} as const;

export interface GenerateRequest {
  text: string;
  voice_id: string;
  mode: string;
  original_text?: string;
  voice_type?: GenerateRequestVoiceType;
  voice_name?: string;
}

export interface GenerateResponse {
  audio_url: string;
  generation_id: number;
  duration_seconds?: number;
}

export type GalleryEntryReactions = { [key: string]: number };

export interface GalleryEntry {
  id: number;
  original_text: string;
  translated_text: string;
  mode: string;
  voice_type?: string;
  voice_name?: string;
  audio_url: string;
  reactions?: GalleryEntryReactions;
  created_at: string;
}

export interface GalleryResponse {
  entries: GalleryEntry[];
  total: number;
}

export interface SaveGalleryRequest {
  generation_id: number;
}

export interface ReactRequest {
  emoji: string;
}

export interface ReactResponse {
  success: boolean;
  emoji: string;
}

export type CloneVoiceBody = {
  audio: Blob;
};

export type GetGalleryParams = {
  limit?: number;
  offset?: number;
};

export type GetMyRantsParams = {
  limit?: number;
  offset?: number;
};