# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DramaGen API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Translate mild text to dramatic text
 */
export const translateTextBodyTextMax = 500;

export const TranslateTextBody = zod.object({
  text: zod.string().max(translateTextBodyTextMax),
  mode: zod.enum([
    "rage",
    "passive_aggressive",
    "disappointed_parent",
    "telenovela",
    "tiktok",
    "drill_sergeant",
    "corporate",
    "ice_cold",
  ]),
});

export const TranslateTextResponse = zod.object({
  original: zod.string(),
  translated: zod.string(),
  mode: zod.string(),
});

/**
 * @summary Get preset character voices
 */
export const GetPresetVoicesResponse = zod.object({
  voices: zod.array(
    zod.object({
      id: zod.string(),
      voice_id: zod.string().optional(),
      name: zod.string(),
      description: zod.string(),
      emoji: zod.string(),
      ready: zod.boolean(),
    }),
  ),
});

/**
 * @summary Clone user voice from audio recording
 */
export const CloneVoiceBody = zod.object({
  audio: zod.instanceof(File),
});

export const CloneVoiceResponse = zod.object({
  voice_id: zod.string(),
  name: zod.string(),
});

/**
 * @summary Generate dramatic audio from text
 */
export const GenerateAudioBody = zod.object({
  text: zod.string(),
  voice_id: zod.string(),
  mode: zod.string(),
  original_text: zod.string().optional(),
  voice_type: zod.enum(["clone", "preset"]).optional(),
  voice_name: zod.string().optional(),
});

export const GenerateAudioResponse = zod.object({
  audio_url: zod.string(),
  generation_id: zod.number(),
  duration_seconds: zod.number().optional(),
});

/**
 * @summary Get recent public gallery entries
 */
export const getGalleryQueryLimitDefault = 20;
export const getGalleryQueryOffsetDefault = 0;

export const GetGalleryQueryParams = zod.object({
  limit: zod.coerce.number().default(getGalleryQueryLimitDefault),
  offset: zod.coerce.number().default(getGalleryQueryOffsetDefault),
});

export const GetGalleryResponse = zod.object({
  entries: zod.array(
    zod.object({
      id: zod.number(),
      original_text: zod.string(),
      translated_text: zod.string(),
      mode: zod.string(),
      voice_type: zod.string().optional(),
      voice_name: zod.string().optional(),
      audio_url: zod.string(),
      reactions: zod.record(zod.string(), zod.number()).optional(),
      created_at: zod.string(),
    }),
  ),
  total: zod.number(),
});

/**
 * @summary Save a generation to the public gallery
 */
export const SaveToGalleryBody = zod.object({
  generation_id: zod.number(),
});

export const SaveToGalleryResponse = zod.object({
  id: zod.number(),
  original_text: zod.string(),
  translated_text: zod.string(),
  mode: zod.string(),
  voice_type: zod.string().optional(),
  voice_name: zod.string().optional(),
  audio_url: zod.string(),
  reactions: zod.record(zod.string(), zod.number()).optional(),
  created_at: zod.string(),
});

/**
 * @summary Get current user's generation history
 */
export const getMyRantsQueryLimitDefault = 20;
export const getMyRantsQueryOffsetDefault = 0;

export const GetMyRantsQueryParams = zod.object({
  limit: zod.coerce.number().default(getMyRantsQueryLimitDefault),
  offset: zod.coerce.number().default(getMyRantsQueryOffsetDefault),
});

export const GetMyRantsResponse = zod.object({
  entries: zod.array(
    zod.object({
      id: zod.number(),
      original_text: zod.string(),
      translated_text: zod.string(),
      mode: zod.string(),
      voice_type: zod.string().optional(),
      voice_name: zod.string().optional(),
      audio_url: zod.string(),
      reactions: zod.record(zod.string(), zod.number()).optional(),
      created_at: zod.string(),
    }),
  ),
  total: zod.number(),
});

/**
 * @summary Add an emoji reaction to a gallery entry
 */
export const ReactToGalleryParams = zod.object({
  id: zod.coerce.number(),
});

export const ReactToGalleryBody = zod.object({
  emoji: zod.string(),
});

export const ReactToGalleryResponse = zod.object({
  success: zod.boolean(),
  emoji: zod.string(),
});